Move Grid background into new directory
